+++ lib/street_address/version.rb
@@ -1,0 +1,3 @@
+module StreetAddress
+  VERSION = '2.0.0'
+end
