+++ street_address.gemspec
@@ -1,11 +1,11 @@
 $LOAD_PATH.unshift 'lib'
 
-require 'street_address'
+require 'street_address/version'
 
 Gem::Specification.new do |s|
   s.name = "StreetAddress"
   s.licenses = ['MIT']
-  s.version = StreetAddress::US::VERSION
+  s.version = StreetAddress::VERSION
   s.date = Time.now.strftime('%Y-%m-%d')
   s.summary = "Parse Addresses into substituent parts. This gem includes US only."
   s.authors = ["Derrek Long"]
